rename data utils to metadata
